add print_wheels.py script + refactor Makefile
