+++ MANIFEST.in
@@ -126,6 +126,7 @@ include scripts/internal/print_access_denied.py
 include scripts/internal/print_announce.py
 include scripts/internal/print_api_speed.py
 include scripts/internal/print_timeline.py
+include scripts/internal/print_wheels.py
 include scripts/internal/purge_installation.py
 include scripts/internal/tidelift.py
 include scripts/internal/winmake.py
