+++ Makefile
@@ -208,6 +208,25 @@ install-git-hooks:  ## Install GIT pre-commit hook.
 	chmod +x .git/hooks/pre-commit
 
 # ===================================================================
+# Wheels
+# ===================================================================
+
+download-wheels-appveyor:  ## Download latest wheels hosted on appveyor.
+	$(PYTHON) scripts/internal/download_wheels_appveyor.py --user giampaolo --project psutil
+
+download-wheels-github:  ## Download latest wheels hosted on github.
+	$(PYTHON) scripts/internal/download_wheels_github.py --user=giampaolo --project=psutil --tokenfile=~/.github.token
+
+download-wheels:  ## Download wheels from github and appveyor
+	rm -rf dist
+	${MAKE} download-wheels-appveyor
+	${MAKE} download-wheels-github
+	${MAKE} print-wheels
+
+print-wheels:  ## Print downloaded wheels
+	$(PYTHON) scripts/internal/print_wheels.py
+
+# ===================================================================
 # Distribution
 # ===================================================================
 
@@ -219,24 +238,11 @@ sdist:  ## Create tar.gz source distribution.
 	${MAKE} generate-manifest
 	$(PYTHON) setup.py sdist
 
-wheel:  ## Generate wheel.
-	$(PYTHON) setup.py bdist_wheel
-
-download-wheels-appveyor:  ## Download latest wheels hosted on appveyor.
-	$(PYTHON) scripts/internal/download_wheels_appveyor.py --user giampaolo --project psutil
-
-download-wheels-github:  ## Download latest wheels hosted on github.
-	$(PYTHON) scripts/internal/download_wheels_github.py --user=giampaolo --project=psutil --tokenfile=~/.github.token
-
-download-wheels:  ## Download wheels from github and appveyor
-	${MAKE} download-wheels-appveyor
-	${MAKE} download-wheels-github
-
 upload-src:  ## Upload source tarball on https://pypi.org/project/psutil/
 	${MAKE} sdist
-	$(PYTHON) setup.py sdist upload
+	$(PYTHON) -m twine upload dist/*.tar.gz
 
-upload-win-wheels:  ## Upload wheels in dist/* directory on PyPI.
+upload-wheels:  ## Upload wheels in dist/* directory on PyPI.
 	$(PYTHON) -m twine upload dist/*.whl
 
 # --- others
@@ -257,7 +263,7 @@ pre-release:  ## Check if we're ready to produce a new release.
 	${MAKE} install
 	${MAKE} generate-manifest
 	git diff MANIFEST.in > /dev/null  # ...otherwise 'git diff-index HEAD' will complain
-	${MAKE} win-download-wheels
+	${MAKE} download-wheels
 	${MAKE} sdist
 	$(PYTHON) -c \
 		"from psutil import __version__ as ver; \
